add objects.py
