+++ objects.py
@@ -1,0 +1,25 @@
+import pygame
+
+def collide(obj1, obj2):
+    offset_x = obj2.x - obj1.x
+    offset_y = obj2.y - obj1.y
+    return obj1.mask.overlap(obj2.mask, (offset_x, offset_y)) != None
+
+class Laser:
+    def __init__(self, x, y, img):
+        self.x = x
+        self.y = y
+        self.img = img
+        self.mask = pygame.mask.from_surface(self.img)
+
+    def draw(self, window):
+        window.blit(self.img, (self.x, self.y))
+    
+    def move(self, laser_vel):
+        self.y += laser_vel
+
+    def off_screen(self, height):
+        return self.y <= 0 or self.y >= height
+
+    def collision(self, obj):
+        return collide(obj, self)
